+++ packages/e2e-tests/selectors.ts
@@ -87,3 +87,15 @@ export class RegisterUsernameModal {
     await t.click(submitButton)
   }
 }
+
+export class DebugModeModal {
+  get title() {
+    return Selector('h3').withText('App is running in debug mode')
+  }
+
+  async close() {
+    if (await this.title.exists) {
+      await t.click(Selector('button').withText('Understand'))
+    }
+  }
+}