+++ packages/frontend/.storybook/preview.js
@@ -1,3 +1,5 @@
+import 'cypress-storybook/react'
+
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
   controls: {