+++ packages/frontend/src/main/main.ts
@@ -5,6 +5,7 @@ import electronLocalshortcut from 'electron-localshortcut'
 import debug from 'debug'
 import path from 'path'
 import url from 'url'
+import fs from 'fs'
 import config from './config'
 import { DataServer, ConnectionsManager } from '@quiet/waggle'
 import { waggleVersion, runWaggle } from './waggleManager'
@@ -19,12 +20,26 @@ const log = Object.assign(debug('frontend:main'), {
   error: debug('frontend:main:err')
 })
 
-electronStore.set('appDataPath', app.getPath('appData'))
-electronStore.set('waggleVersion', waggleVersion)
-
 export const isDev = process.env.NODE_ENV === 'development'
 export const isE2Etest = process.env.E2E_TEST === 'true'
 const webcrypto = new Crypto()
+
+if (isDev || process.env.DATA_DIR) {
+  const dataDir = process.env.DATA_DIR || 'Quietdev'
+  const appDataPath = path.join(app.getPath('appData'), dataDir)
+
+  if (!fs.existsSync(appDataPath)) {
+    fs.mkdirSync(appDataPath)
+  }
+
+  const newUserDataPath = path.join(appDataPath, 'Quiet')
+
+  app.setPath('appData', appDataPath)
+  app.setPath('userData', newUserDataPath)
+
+  electronStore.set('appDataPath', app.getPath('appData'))
+  electronStore.set('waggleVersion', waggleVersion)
+}
 
 interface IWindowSize {
   width: number
@@ -322,11 +337,7 @@ app.on('ready', async () => {
 
 app.setAsDefaultProtocolClient('quiet')
 
-app.on('before-quit', async e => {
-  if (waggleProcess !== null) {
-    await waggleProcess.connectionsManager.closeAllServices()
-    await waggleProcess.dataServer.close()
-  }
+app.on('before-quit', () => {
   if (browserWidth && browserHeight) {
     electronStore.set('windowSize', {
       width: browserWidth,
@@ -336,7 +347,11 @@ app.on('before-quit', async e => {
 })
 
 // Quit when all windows are closed.
-app.on('window-all-closed', () => {
+app.on('window-all-closed', async () => {
+  if (waggleProcess !== null) {
+    await waggleProcess.connectionsManager.closeAllServices()
+    await waggleProcess.dataServer.close()
+  }
   // On macOS it is common for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
   // NOTE: temporarly quit macos when using 'X'. Reloading the app loses the connection with waggle. To be fixed.