+++ packages/frontend/src/renderer/containers/widgets/settings/SettingsModal.tsx
@@ -8,7 +8,9 @@ import { communities, identity } from '@quiet/nectar'
 const SettingsModalContainer = () => {
   const modal = useModal(ModalName.accountSettingsModal)
 
-  const owner = useSelector(communities.selectors.isOwner)
+  const user = useSelector(identity.selectors.currentIdentity)?.nickname || 'Settings'
+  const community = useSelector(communities.selectors.currentCommunity)
+  const owner = Boolean(community?.CA)
 
   return <SettingsModal title={'Settings'} owner={owner} {...modal} />
 }