+++ packages/frontend/src/rtl-tests/community.join.test.tsx
@@ -18,15 +18,15 @@ import { socketEventData } from '../renderer/testUtils/socket'
 import {
   identity,
   communities,
-  getFactory,
   SocketActionTypes,
   RegisterUserCertificatePayload,
   InitCommunityPayload,
-  Identity,
   Community,
   createUserCertificateTestHelper,
   ErrorCodes,
-  ErrorMessages
+  ErrorMessages,
+  getFactory,
+  errors
 } from '@quiet/nectar'
 import Channel from '../renderer/components/Channel/Channel'
 
@@ -41,9 +41,6 @@ describe('User', () => {
   })
 
   it('joins community and registers username', async () => {
-    let community: Community
-    let alice: Identity
-
     const { store, runSaga } = await prepareStore(
       {},
       socket // Fork Nectar's sagas
@@ -66,50 +63,53 @@ describe('User', () => {
       .spyOn(socket, 'emit')
       .mockImplementation(async (action: SocketActionTypes, ...input: any[]) => {
         if (action === SocketActionTypes.CREATE_NETWORK) {
-          const data = input as socketEventData<[string]>
-          community = (
-            await factory.build<typeof communities.actions.addNewCommunity>('Community', {
-              id: data[0]
-            })
-          ).payload
-          alice = (
-            await factory.build<typeof identity.actions.addNewIdentity>('Identity', {
-              id: community.id,
-              nickname: 'alice'
-            })
-          ).payload
-
+          const data = input as socketEventData<[Community]>
+          const payload = data[0]
           return socket.socketClient.emit(SocketActionTypes.NETWORK, {
-            id: community.id,
-            payload: {
-              hiddenService: alice.hiddenService,
-              peerId: alice.peerId
+            community: payload,
+            network: {
+              hiddenService: {
+                onionAddress: 'onionAddress',
+                privKey: 'privKey'
+              },
+              peerId: {
+                id: 'peerId'
+              }
             }
           })
         }
         if (action === SocketActionTypes.REGISTER_USER_CERTIFICATE) {
           const data = input as socketEventData<[RegisterUserCertificatePayload]>
           const payload = data[0]
-          expect(payload.id).toEqual(community.id)
-          const certificate = await createUserCertificateTestHelper(
+          const user = identity.selectors.currentIdentity(store.getState())
+          // This community serves only as a mocked object for generating valid crytpo data (certificate, rootCA)
+          const communityHelper = (
+            await factory.build<typeof communities.actions.addNewCommunity>('Community', {
+              id: data[0]
+            })
+          ).payload
+          const certificateHelper = await createUserCertificateTestHelper(
             {
-              nickname: alice.nickname,
-              commonName: alice.hiddenService.onionAddress,
-              peerId: alice.peerId.id
+              nickname: user.nickname,
+              commonName: communityHelper.registrarUrl,
+              peerId: user.peerId.id
             },
-            community.CA
+            communityHelper.CA
           )
+          const certificate = certificateHelper.userCert.userCertObject.certificate
+          const rootCa = communityHelper.CA.rootCertString
           return socket.socketClient.emit(SocketActionTypes.SEND_USER_CERTIFICATE, {
-            id: payload.id,
+            communityId: payload.communityId,
             payload: {
               certificate: certificate,
-              rootCa: community.CA.rootCertString
+              rootCa: rootCa
             }
           })
         }
         if (action === SocketActionTypes.LAUNCH_COMMUNITY) {
           const data = input as socketEventData<[InitCommunityPayload]>
           const payload = data[0]
+          const community = communities.selectors.currentCommunity(store.getState())
           expect(payload.id).toEqual(community.id)
           socket.socketClient.emit(SocketActionTypes.COMMUNITY, {
             id: payload.id
@@ -160,7 +160,7 @@ describe('User', () => {
     userEvent.click(createUsernameButton)
 
     // Wait for the actions that updates the store
-    await act(async () => {})
+    await act(async () => { })
 
     // Check if join/username modals are gone
     expect(joinCommunityTitle).not.toBeVisible()
@@ -172,20 +172,19 @@ describe('User', () => {
 
     expect(actions).toMatchInlineSnapshot(`
       Array [
-        "Modals/openModal",
-        "Modals/openModal",
-        "Communities/joinCommunity",
+        "Communities/createNetwork",
+        "Communities/responseCreateNetwork",
         "Communities/addNewCommunity",
-        "PublicChannels/addPublicChannelsList",
         "Communities/setCurrentCommunity",
-        "Communities/responseCreateCommunity",
+        "PublicChannels/addPublicChannelsList",
         "Identity/addNewIdentity",
+        "Modals/closeModal",
+        "Modals/openModal",
         "Identity/registerUsername",
-        "Identity/updateUsername",
-        "Identity/createUserCsr",
         "Identity/registerCertificate",
         "Communities/storePeerList",
         "Identity/storeUserCertificate",
+        "Modals/closeModal",
         "Communities/updateCommunity",
         "Communities/updateCommunityData",
         "Communities/launchCommunity",
@@ -195,18 +194,12 @@ describe('User', () => {
         "PublicChannels/subscribeToAllTopics",
         "PublicChannels/subscribeToTopic",
         "PublicChannels/addChannel",
-        "Modals/closeModal",
-        "Modals/closeModal",
-        "Modals/closeModal",
-        "Modals/closeModal",
+        "PublicChannels/setChannelLoadingSlice",
       ]
     `)
   })
 
-  it('sees proper registration error when trying to join with already used username', async () => {
-    let community: Community
-    let alice: Identity
-
+  it('sees proper registration error when trying to join with already taken username', async () => {
     const { store, runSaga } = await prepareStore(
       {},
       socket // Fork Nectar's sagas
@@ -224,61 +217,35 @@ describe('User', () => {
       store
     )
 
-    const factory = await getFactory(store)
-
     jest
       .spyOn(socket, 'emit')
       .mockImplementation(async (action: SocketActionTypes, ...input: any[]) => {
         if (action === SocketActionTypes.CREATE_NETWORK) {
-          const data = input as socketEventData<[string]>
-          community = (
-            await factory.build<typeof communities.actions.addNewCommunity>('Community', {
-              id: data[0]
-            })
-          ).payload
-          alice = (
-            await factory.build<typeof identity.actions.addNewIdentity>('Identity', {
-              id: community.id,
-              nickname: 'alice'
-            })
-          ).payload
+          const data = input as socketEventData<[Community]>
+          const payload = data[0]
           return socket.socketClient.emit(SocketActionTypes.NETWORK, {
-            id: community.id,
-            payload: {
-              hiddenService: alice.hiddenService,
-              peerId: alice.peerId
+            community: payload,
+            network: {
+              hiddenService: {
+                onionAddress: 'onionAddress',
+                privKey: 'privKey'
+              },
+              peerId: {
+                id: 'peerId'
+              }
             }
           })
         }
         if (action === SocketActionTypes.REGISTER_USER_CERTIFICATE) {
           const data = input as socketEventData<[RegisterUserCertificatePayload]>
           const payload = data[0]
-          expect(payload.id).toEqual(community.id)
+          const community = communities.selectors.currentCommunity(store.getState())
+          expect(payload.communityId).toEqual(community.id)
           socket.socketClient.emit(SocketActionTypes.ERROR, {
             type: SocketActionTypes.REGISTRAR,
+            code: ErrorCodes.FORBIDDEN,
             message: ErrorMessages.USERNAME_TAKEN,
-            code: ErrorCodes.VALIDATION,
             community: community.id
-          })
-        }
-        if (action === SocketActionTypes.LAUNCH_COMMUNITY) {
-          const data = input as socketEventData<[InitCommunityPayload]>
-          const payload = data[0]
-          expect(payload.id).toEqual(community.id)
-          socket.socketClient.emit(SocketActionTypes.COMMUNITY, {
-            id: payload.id
-          })
-          socket.socketClient.emit(SocketActionTypes.RESPONSE_GET_PUBLIC_CHANNELS, {
-            communityId: community.id,
-            channels: {
-              general: {
-                name: 'general',
-                description: 'string',
-                owner: 'owner',
-                timestamp: 0,
-                address: 'general'
-              }
-            }
           })
         }
       })
@@ -314,7 +281,7 @@ describe('User', () => {
     userEvent.click(createUsernameButton)
 
     // Wait for the actions that updates the store
-    await act(async () => {})
+    await act(async () => { })
 
     // Check if 'username taken' error message is visible
     expect(createUsernameTitle).toBeVisible()
@@ -323,20 +290,126 @@ describe('User', () => {
 
     expect(actions).toMatchInlineSnapshot(`
       Array [
-        "Modals/openModal",
-        "Modals/openModal",
-        "Communities/joinCommunity",
+        "Communities/createNetwork",
+        "Communities/responseCreateNetwork",
         "Communities/addNewCommunity",
-        "PublicChannels/addPublicChannelsList",
         "Communities/setCurrentCommunity",
-        "Communities/responseCreateCommunity",
+        "PublicChannels/addPublicChannelsList",
         "Identity/addNewIdentity",
+        "Modals/closeModal",
+        "Modals/openModal",
         "Identity/registerUsername",
-        "Identity/updateUsername",
-        "Identity/createUserCsr",
         "Identity/registerCertificate",
         "Errors/addError",
+      ]
+    `)
+  })
+
+  it('clears error before sending another username registration request', async () => {
+    const { store, runSaga } = await prepareStore(
+      {},
+      socket // Fork Nectar's sagas
+    )
+
+    store.dispatch(modalsActions.openModal({ name: ModalName.joinCommunityModal }))
+
+    renderComponent(
+      <>
+        <LoadingPanel />
+        <JoinCommunity />
+        <CreateUsername />
+        <Channel />
+      </>,
+      store
+    )
+
+    jest
+      .spyOn(socket, 'emit')
+      .mockImplementation(async (action: SocketActionTypes, ...input: any[]) => {
+        if (action === SocketActionTypes.CREATE_NETWORK) {
+          const data = input as socketEventData<[Community]>
+          const payload = data[0]
+          return socket.socketClient.emit(SocketActionTypes.NETWORK, {
+            community: payload,
+            network: {
+              hiddenService: {
+                onionAddress: 'onionAddress',
+                privKey: 'privKey'
+              },
+              peerId: {
+                id: 'peerId'
+              }
+            }
+          })
+        }
+      })
+
+    // Log all the dispatched actions in order
+    const actions = []
+    runSaga(function* (): Generator {
+      while (true) {
+        const action = yield* take()
+        actions.push(action.type)
+      }
+    })
+
+    // Confirm proper modal title is displayed
+    const dictionary = JoinCommunityDictionary()
+    const joinCommunityTitle = screen.getByText(dictionary.header)
+    expect(joinCommunityTitle).toBeVisible()
+
+    // Enter community address and hit button
+    const joinCommunityInput = screen.getByPlaceholderText(dictionary.placeholder)
+    const joinCommunityButton = screen.getByText(dictionary.button)
+    userEvent.type(joinCommunityInput, '3lyn5yjwwb74he5olv43eej7knt34folvrgrfsw6vzitvkxmc5wpe4yd')
+    userEvent.click(joinCommunityButton)
+
+    // Confirm user is being redirected to username registration
+    const createUsernameTitle = await screen.findByText('Register a username')
+    expect(createUsernameTitle).toBeVisible()
+
+    await act(async () => {
+      const community = communities.selectors.currentCommunity(store.getState())
+      store.dispatch(
+        errors.actions.addError({
+          type: SocketActionTypes.REGISTRAR,
+          code: ErrorCodes.FORBIDDEN,
+          message: ErrorMessages.USERNAME_TAKEN,
+          community: community.id
+        })
+      )
+    })
+
+    // Check if 'username taken' error message is visible
+    expect(createUsernameTitle).toBeVisible()
+    expect(await screen.findByText(ErrorMessages.USERNAME_TAKEN)).toBeVisible()
+
+    // Enter username and hit button
+    const createUsernameInput = screen.getByPlaceholderText('Enter a username')
+    const createUsernameButton = screen.getByText('Register')
+    userEvent.type(createUsernameInput, 'bob')
+    userEvent.click(createUsernameButton)
+
+    // Wait for the actions that updates the store
+    await act(async () => { })
+
+    // Check if 'username taken' error message disappeared
+    expect(await screen.queryByText(ErrorMessages.USERNAME_TAKEN)).toBeNull()
+
+    expect(actions).toMatchInlineSnapshot(`
+      Array [
+        "Communities/createNetwork",
+        "Communities/responseCreateNetwork",
+        "Communities/addNewCommunity",
+        "Communities/setCurrentCommunity",
+        "PublicChannels/addPublicChannelsList",
+        "Identity/addNewIdentity",
         "Modals/closeModal",
+        "Modals/openModal",
+        "Errors/addError",
+        "Errors/clearError",
+        "Identity/registerUsername",
+        "Identity/registerCertificate",
       ]
     `)
   })