+++ packages/nectar/src/sagas/errors/handleErrors/handleErrors.saga.ts
@@ -1,25 +1,22 @@
 import { PayloadAction } from '@reduxjs/toolkit'
+import { call, delay, put, select } from 'typed-redux-saga'
 import { identitySelectors } from '../../identity/identity.selectors'
 import { identityActions } from '../../identity/identity.slice'
-import { call, put, select } from 'typed-redux-saga'
 import { errorsActions } from '../errors.slice'
-import { communitiesSelectors } from '../../communities/communities.selectors'
 import { SocketActionTypes } from '../../socket/const/actionTypes'
-import { ErrorPayload } from '../errors.types'
+import { ErrorCodes, ErrorPayload } from '../errors.types'
+import { RegisterCertificatePayload } from '../../identity/identity.types'
 import logger from '../../../utils/logger'
 
 const log = logger('errors')
 
 function* retryRegistration(communityId: string) {
   const identity = yield* select(identitySelectors.selectById(communityId))
-  const registrarAddress = yield* select(
-    communitiesSelectors.registrarUrl(communityId)
-  )
 
-  const payload = {
-    communityId,
-    userCsr: identity.userCsr,
-    registrarAddress
+  const payload: RegisterCertificatePayload = {
+    communityId: communityId,
+    nickname: identity.nickname,
+    userCsr: identity.userCsr
   }
 
   yield* put(identityActions.registerCertificate(payload))
@@ -31,7 +28,8 @@ export function* handleErrorsSaga(
 ): Generator {
   const error: ErrorPayload = action.payload
   if (error.type === SocketActionTypes.REGISTRAR) {
-    if (error.code === 500) {
+    if (error.code === ErrorCodes.NOT_FOUND) {
+      yield* call(delay, 15000)
       yield* call(retryRegistration, error.community)
     }
   }