+++ packages/frontend/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0-alpha.3](https://github.com/ZbayApp/monorepo/compare/quiet@0.4.0-alpha.2...quiet@0.4.0-alpha.3) (2022-03-11)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [0.4.0-alpha.2](https://github.com/ZbayApp/monorepo/compare/quiet@0.4.0-alpha.0...quiet@0.4.0-alpha.2) (2022-03-11)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [0.4.0-alpha.1](https://github.com/ZbayApp/monorepo/compare/quiet@0.4.0-alpha.0...quiet@0.4.0-alpha.1) (2022-03-11)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [0.4.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@0.3.0-alpha.7...quiet@0.4.0-alpha.0) (2022-03-08)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 # [0.3.0-alpha.8](https://github.com/ZbayApp/monorepo/compare/quiet@0.3.0-alpha.7...quiet@0.3.0-alpha.8) (2022-02-28)
 
 **Note:** Version bump only for package quiet